xss_tester/requirements.txt
# HTTP client
httpx>=0.27.0

# HTML parsing
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Data models & validation
pydantic>=2.6.0

# Browser automation (XSS validation)
playwright>=1.43.0

# YAML payloads
PyYAML>=6.0.1

# barra progrés
tqdm